Create insert element in an array
Given a sorted list with an unsorted number e in the rightmost cell, simple code to insert e into the array so that it remains sorted

Print the array every time a value is shifted in the array until the array is fully sorted. The goal of this challenge is to follow the correct order of insertion sort.